server/requirements/base.txt
django==3.2.8
django-environ==0.4.5
django-crispy-forms==1.13.0
djangorestframework==3.13.1
django-filter==21.1
django-cors-headers==3.4.0
django-allauth==0.47.0
django-recaptcha>=1.2.0
django-feature-policy==3.4.0
pytz==2018.7
redis==2.10.6
hiredis==1.1.0
django-redis==4.4.4
boto3>=1.7.19



